+++ R/data.R
@@ -38,7 +38,9 @@
 #'   \item{resp}{boolean indicating whether the firm did respond to the survey or not}
 #'   \item{hrg}{homogeneous response group used for the unit non-response correction}
 #'   \item{response_prob_est}{response probability of the unit estimated using homogeneous response groups}
-#'   \item{w_nr}{weight after unit non-response correction}
+#'   \item{w_nrc}{weight after unit non-response correction}
+#'   \item{N_58, N_59, N_60, N_61, N_62, N_63, turnover_58, turnover_59, turnover_60, turnover_61, turnover_62, turnover_63}{calibration variables (number of firms and turnover broken down by economic sub-sector)}
+#'   \item{w_calib}{calibrated weight}
 #' }
 #' 
 #' @seealso \code{\link{ict_pop}} \code{\link{ict_survey}}
@@ -57,9 +59,6 @@
 #'   \item{division}{identifier of the economic sub-sector}
 #'   \item{employees}{number of employees}
 #'   \item{turnover}{firm turnover, in euros}
-#'   \item{w_sample}{sampling weight}
-#'   \item{w_nr}{weight after unit non-response correction}
-#'   \item{N_58, N_59, N_60, N_61, N_62, N_63, turnover_58, turnover_59, turnover_60, turnover_61, turnover_62, turnover_63}{calibration variables (number of firms and turnover broken down by economic sub-sector)}
 #'   \item{w_calib}{calibrated weight}
 #'   \item{speed_quanti, speed_quanti_NA}{internet connection speed of the firm in Mbits, without or with missing values}
 #'   \item{speed_quali, speed_quali_NA}{internet connection speed of the firm recoded in classes, without or with missing values}